@@ -4,48 +4,31 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.common.MinecraftForge;
 
-import net.minecraft.world.IWorld;
-import net.minecraft.world.GameType;
-import net.minecraft.item.ItemStack;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.Entity;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.level.GameType;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.server.level.ServerPlayer;
 
-import java.util.Map;
+import io.itch.awesomekalin.noob.init.NoobModItems;
 
-import io.itch.awesomekalin.noob.item.CreativeGauntletItem;
-import io.itch.awesomekalin.noob.NoobModElements;
-import io.itch.awesomekalin.noob.NoobMod;
-
-@NoobModElements.ModElement.Tag
-public class CreativeGauntletRightClickedProcedure extends NoobModElements.ModElement {
-	public CreativeGauntletRightClickedProcedure(NoobModElements instance) {
-		super(instance, 52);
-	}
-
-	public static void executeProcedure(Map<String, Object> dependencies) {
-		if (dependencies.get("entity") == null) {
-			if (!dependencies.containsKey("entity"))
-				NoobMod.LOGGER.warn("Failed to load dependency entity for procedure CreativeGauntletRightClicked!");
+public class CreativeGauntletRightClickedProcedure {
+	public static void execute(LevelAccessor world, Entity entity) {
+		if (entity == null)
 			return;
-		}
-		if (dependencies.get("world") == null) {
-			if (!dependencies.containsKey("world"))
-				NoobMod.LOGGER.warn("Failed to load dependency world for procedure CreativeGauntletRightClicked!");
-			return;
-		}
-		Entity entity = (Entity) dependencies.get("entity");
-		IWorld world = (IWorld) dependencies.get("world");
-		if (entity instanceof PlayerEntity)
-			((PlayerEntity) entity).setGameType(GameType.CREATIVE);
-		if (entity instanceof PlayerEntity) {
-			ItemStack _stktoremove = new ItemStack(CreativeGauntletItem.block);
-			((PlayerEntity) entity).inventory.clearMatchingItems(p -> _stktoremove.getItem() == p.getItem(), (int) 1);
+		if (entity instanceof ServerPlayer _player)
+			_player.setGameMode(GameType.CREATIVE);
+		if (entity instanceof Player _player) {
+			ItemStack _stktoremove = new ItemStack(NoobModItems.CREATIVE_GAUNTLET.get());
+			_player.getInventory().clearOrCountMatchingItems(p -> _stktoremove.getItem() == p.getItem(), 1, _player.inventoryMenu.getCraftSlots());
 		}
 		new Object() {
 			private int ticks = 0;
 			private float waitTicks;
-			private IWorld world;
-			public void start(IWorld world, int waitTicks) {
+			private LevelAccessor world;
+
+			public void start(LevelAccessor world, int waitTicks) {
 				this.waitTicks = waitTicks;
 				MinecraftForge.EVENT_BUS.register(this);
 				this.world = world;
@@ -61,10 +44,10 @@ public class CreativeGauntletRightClickedProcedure extends NoobModElements.ModEl
 			}
 
 			private void run() {
-				if (entity instanceof PlayerEntity)
-					((PlayerEntity) entity).setGameType(GameType.SURVIVAL);
+				if (entity instanceof ServerPlayer _player)
+					_player.setGameMode(GameType.SURVIVAL);
 				MinecraftForge.EVENT_BUS.unregister(this);
 			}
-		}.start(world, (int) 600);
+		}.start(world, 600);
 	}
 }
