@@ -1,77 +1,39 @@
 package io.itch.awesomekalin.noob.procedures;
 
-import net.minecraftforge.fml.network.NetworkHooks;
+import net.minecraftforge.network.NetworkHooks;
 
-import net.minecraft.world.IWorld;
-import net.minecraft.util.text.StringTextComponent;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.network.PacketBuffer;
-import net.minecraft.inventory.container.INamedContainerProvider;
-import net.minecraft.inventory.container.Container;
-import net.minecraft.entity.player.ServerPlayerEntity;
-import net.minecraft.entity.player.PlayerInventory;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.Entity;
-
-import java.util.Map;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.inventory.AbstractContainerMenu;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.core.BlockPos;
 
 import io.netty.buffer.Unpooled;
 
-import io.itch.awesomekalin.noob.gui.NoobChestGUIGui;
-import io.itch.awesomekalin.noob.NoobModElements;
-import io.itch.awesomekalin.noob.NoobMod;
+import io.itch.awesomekalin.noob.world.inventory.NoobChestGUIMenu;
 
-@NoobModElements.ModElement.Tag
-public class NoobChestOnBlockRightClickedProcedure extends NoobModElements.ModElement {
-	public NoobChestOnBlockRightClickedProcedure(NoobModElements instance) {
-		super(instance, 3);
-	}
-
-	public static void executeProcedure(Map<String, Object> dependencies) {
-		if (dependencies.get("entity") == null) {
-			if (!dependencies.containsKey("entity"))
-				NoobMod.LOGGER.warn("Failed to load dependency entity for procedure NoobChestOnBlockRightClicked!");
+public class NoobChestOnBlockRightClickedProcedure {
+	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
+		if (entity == null)
 			return;
-		}
-		if (dependencies.get("x") == null) {
-			if (!dependencies.containsKey("x"))
-				NoobMod.LOGGER.warn("Failed to load dependency x for procedure NoobChestOnBlockRightClicked!");
-			return;
-		}
-		if (dependencies.get("y") == null) {
-			if (!dependencies.containsKey("y"))
-				NoobMod.LOGGER.warn("Failed to load dependency y for procedure NoobChestOnBlockRightClicked!");
-			return;
-		}
-		if (dependencies.get("z") == null) {
-			if (!dependencies.containsKey("z"))
-				NoobMod.LOGGER.warn("Failed to load dependency z for procedure NoobChestOnBlockRightClicked!");
-			return;
-		}
-		if (dependencies.get("world") == null) {
-			if (!dependencies.containsKey("world"))
-				NoobMod.LOGGER.warn("Failed to load dependency world for procedure NoobChestOnBlockRightClicked!");
-			return;
-		}
-		Entity entity = (Entity) dependencies.get("entity");
-		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
-		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
-		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		IWorld world = (IWorld) dependencies.get("world");
 		{
-			Entity _ent = entity;
-			if (_ent instanceof ServerPlayerEntity) {
-				BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
-				NetworkHooks.openGui((ServerPlayerEntity) _ent, new INamedContainerProvider() {
+			if (entity instanceof ServerPlayer _ent) {
+				BlockPos _bpos = new BlockPos(x, y, z);
+				NetworkHooks.openGui((ServerPlayer) _ent, new MenuProvider() {
 					@Override
-					public ITextComponent getDisplayName() {
-						return new StringTextComponent("NoobChestGUI");
+					public Component getDisplayName() {
+						return new TextComponent("NoobChestGUI");
 					}
 
 					@Override
-					public Container createMenu(int id, PlayerInventory inventory, PlayerEntity player) {
-						return new NoobChestGUIGui.GuiContainerMod(id, inventory, new PacketBuffer(Unpooled.buffer()).writeBlockPos(_bpos));
+					public AbstractContainerMenu createMenu(int id, Inventory inventory, Player player) {
+						return new NoobChestGUIMenu(id, inventory, new FriendlyByteBuf(Unpooled.buffer()).writeBlockPos(_bpos));
 					}
 				}, _bpos);
 			}
