@@ -1,56 +1,25 @@
 package io.itch.awesomekalin.noob.procedures;
 
-import net.minecraft.world.server.ServerWorld;
-import net.minecraft.world.IWorld;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.effect.LightningBoltEntity;
-import net.minecraft.entity.Entity;
+import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.level.LevelAccessor;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.LightningBolt;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.core.BlockPos;
 
-import java.util.Map;
-
-import io.itch.awesomekalin.noob.NoobModElements;
-import io.itch.awesomekalin.noob.NoobMod;
-
-@NoobModElements.ModElement.Tag
-public class NoobFoodFoodEatenProcedure extends NoobModElements.ModElement {
-	public NoobFoodFoodEatenProcedure(NoobModElements instance) {
-		super(instance, 8);
-	}
-
-	public static void executeProcedure(Map<String, Object> dependencies) {
-		if (dependencies.get("entity") == null) {
-			if (!dependencies.containsKey("entity"))
-				NoobMod.LOGGER.warn("Failed to load dependency entity for procedure NoobFoodFoodEaten!");
+public class NoobFoodFoodEatenProcedure {
+	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
+		if (entity == null)
 			return;
+		if (world instanceof ServerLevel _level) {
+			LightningBolt entityToSpawn = EntityType.LIGHTNING_BOLT.create(_level);
+			entityToSpawn.moveTo(Vec3.atBottomCenterOf(new BlockPos(x, y, z)));
+			entityToSpawn.setVisualOnly(false);
+			_level.addFreshEntity(entityToSpawn);
 		}
-		if (dependencies.get("x") == null) {
-			if (!dependencies.containsKey("x"))
-				NoobMod.LOGGER.warn("Failed to load dependency x for procedure NoobFoodFoodEaten!");
-			return;
-		}
-		if (dependencies.get("y") == null) {
-			if (!dependencies.containsKey("y"))
-				NoobMod.LOGGER.warn("Failed to load dependency y for procedure NoobFoodFoodEaten!");
-			return;
-		}
-		if (dependencies.get("z") == null) {
-			if (!dependencies.containsKey("z"))
-				NoobMod.LOGGER.warn("Failed to load dependency z for procedure NoobFoodFoodEaten!");
-			return;
-		}
-		if (dependencies.get("world") == null) {
-			if (!dependencies.containsKey("world"))
-				NoobMod.LOGGER.warn("Failed to load dependency world for procedure NoobFoodFoodEaten!");
-			return;
-		}
-		Entity entity = (Entity) dependencies.get("entity");
-		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
-		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
-		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		IWorld world = (IWorld) dependencies.get("world");
-		if (world instanceof ServerWorld)
-			((ServerWorld) world).addLightningBolt(new LightningBoltEntity(world.getWorld(), (int) x, (int) y, (int) z, false));
-		if (entity instanceof PlayerEntity)
-			((PlayerEntity) entity).getFoodStats().setFoodLevel((int) 1);
+		if (entity instanceof Player _player)
+			_player.getFoodData().setFoodLevel(1);
 	}
 }
