@@ -1,93 +1,107 @@
 
 package io.itch.awesomekalin.noob.item;
 
-import net.minecraftforge.registries.ObjectHolder;
 import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.api.distmarker.Dist;
 
-import net.minecraft.util.ResourceLocation;
-import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.IArmorMaterial;
-import net.minecraft.item.ArmorItem;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.entity.Entity;
-import net.minecraft.block.Blocks;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.sounds.SoundEvent;
+import net.minecraft.resources.ResourceLocation;
 
-import io.itch.awesomekalin.noob.itemgroup.NoobTabItemGroup;
-import io.itch.awesomekalin.noob.NoobModElements;
+import io.itch.awesomekalin.noob.init.NoobModTabs;
 
-@NoobModElements.ModElement.Tag
-public class DirtArmourArmorItem extends NoobModElements.ModElement {
-	@ObjectHolder("noob:dirt_armour_armor_helmet")
-	public static final Item helmet = null;
-	@ObjectHolder("noob:dirt_armour_armor_chestplate")
-	public static final Item body = null;
-	@ObjectHolder("noob:dirt_armour_armor_leggings")
-	public static final Item legs = null;
-	@ObjectHolder("noob:dirt_armour_armor_boots")
-	public static final Item boots = null;
-	public DirtArmourArmorItem(NoobModElements instance) {
-		super(instance, 45);
-	}
-
-	@Override
-	public void initElements() {
-		IArmorMaterial armormaterial = new IArmorMaterial() {
-			public int getDurability(EquipmentSlotType slot) {
+public abstract class DirtArmourArmorItem extends ArmorItem {
+	public DirtArmourArmorItem(EquipmentSlot slot, Item.Properties properties) {
+		super(new ArmorMaterial() {
+			@Override
+			public int getDurabilityForSlot(EquipmentSlot slot) {
 				return new int[]{13, 15, 16, 11}[slot.getIndex()] * 2;
 			}
 
-			public int getDamageReductionAmount(EquipmentSlotType slot) {
+			@Override
+			public int getDefenseForSlot(EquipmentSlot slot) {
 				return new int[]{0, 1, 1, 0}[slot.getIndex()];
 			}
 
-			public int getEnchantability() {
+			@Override
+			public int getEnchantmentValue() {
 				return 1;
 			}
 
-			public net.minecraft.util.SoundEvent getSoundEvent() {
-				return (net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(""));
+			@Override
+			public SoundEvent getEquipSound() {
+				return ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation(""));
 			}
 
-			public Ingredient getRepairMaterial() {
-				return Ingredient.fromStacks(new ItemStack(Blocks.DIRT));
+			@Override
+			public Ingredient getRepairIngredient() {
+				return Ingredient.of(new ItemStack(Blocks.DIRT));
 			}
 
-			@OnlyIn(Dist.CLIENT)
+			@Override
 			public String getName() {
 				return "dirt_armour_armor";
 			}
 
+			@Override
 			public float getToughness() {
 				return 0f;
 			}
-		};
-		elements.items.add(() -> new ArmorItem(armormaterial, EquipmentSlotType.HEAD, new Item.Properties().group(NoobTabItemGroup.tab)) {
+
 			@Override
-			public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
-				return "noob:textures/models/armor/dirtarmour__layer_" + (slot == EquipmentSlotType.LEGS ? "2" : "1") + ".png";
+			public float getKnockbackResistance() {
+				return 0f;
 			}
-		}.setRegistryName("dirt_armour_armor_helmet"));
-		elements.items.add(() -> new ArmorItem(armormaterial, EquipmentSlotType.CHEST, new Item.Properties().group(NoobTabItemGroup.tab)) {
-			@Override
-			public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
-				return "noob:textures/models/armor/dirtarmour__layer_" + (slot == EquipmentSlotType.LEGS ? "2" : "1") + ".png";
-			}
-		}.setRegistryName("dirt_armour_armor_chestplate"));
-		elements.items.add(() -> new ArmorItem(armormaterial, EquipmentSlotType.LEGS, new Item.Properties().group(NoobTabItemGroup.tab)) {
-			@Override
-			public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
-				return "noob:textures/models/armor/dirtarmour__layer_" + (slot == EquipmentSlotType.LEGS ? "2" : "1") + ".png";
-			}
-		}.setRegistryName("dirt_armour_armor_leggings"));
-		elements.items.add(() -> new ArmorItem(armormaterial, EquipmentSlotType.FEET, new Item.Properties().group(NoobTabItemGroup.tab)) {
-			@Override
-			public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlotType slot, String type) {
-				return "noob:textures/models/armor/dirtarmour__layer_" + (slot == EquipmentSlotType.LEGS ? "2" : "1") + ".png";
-			}
-		}.setRegistryName("dirt_armour_armor_boots"));
+		}, slot, properties);
+	}
+
+	public static class Helmet extends DirtArmourArmorItem {
+		public Helmet() {
+			super(EquipmentSlot.HEAD, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
+		}
+
+		@Override
+		public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
+			return "noob:textures/models/armor/dirtarmour__layer_1.png";
+		}
+	}
+
+	public static class Chestplate extends DirtArmourArmorItem {
+		public Chestplate() {
+			super(EquipmentSlot.CHEST, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
+		}
+
+		@Override
+		public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
+			return "noob:textures/models/armor/dirtarmour__layer_1.png";
+		}
+	}
+
+	public static class Leggings extends DirtArmourArmorItem {
+		public Leggings() {
+			super(EquipmentSlot.LEGS, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
+		}
+
+		@Override
+		public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
+			return "noob:textures/models/armor/dirtarmour__layer_2.png";
+		}
+	}
+
+	public static class Boots extends DirtArmourArmorItem {
+		public Boots() {
+			super(EquipmentSlot.FEET, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
+		}
+
+		@Override
+		public String getArmorTexture(ItemStack stack, Entity entity, EquipmentSlot slot, String type) {
+			return "noob:textures/models/armor/dirtarmour__layer_1.png";
+		}
 	}
 }
