@@ -1,73 +1,49 @@
 
 package io.itch.awesomekalin.noob.item;
 
-import net.minecraftforge.registries.ObjectHolder;
-
-import net.minecraft.world.World;
-import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.item.SwordItem;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.IItemTier;
-import net.minecraft.entity.LivingEntity;
-
-import java.util.Map;
-import java.util.HashMap;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.Tier;
+import net.minecraft.world.item.SwordItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.entity.LivingEntity;
 
 import io.itch.awesomekalin.noob.procedures.MudAxeMobIsHitWithToolProcedure;
-import io.itch.awesomekalin.noob.itemgroup.NoobTabItemGroup;
-import io.itch.awesomekalin.noob.NoobModElements;
+import io.itch.awesomekalin.noob.init.NoobModTabs;
 
-@NoobModElements.ModElement.Tag
-public class MudSwordItem extends NoobModElements.ModElement {
-	@ObjectHolder("noob:mud_sword")
-	public static final Item block = null;
-	public MudSwordItem(NoobModElements instance) {
-		super(instance, 11);
-	}
-
-	@Override
-	public void initElements() {
-		elements.items.add(() -> new SwordItem(new IItemTier() {
-			public int getMaxUses() {
+public class MudSwordItem extends SwordItem {
+	public MudSwordItem() {
+		super(new Tier() {
+			public int getUses() {
 				return 10;
 			}
 
-			public float getEfficiency() {
+			public float getSpeed() {
 				return 1f;
 			}
 
-			public float getAttackDamage() {
+			public float getAttackDamageBonus() {
 				return -1.8f;
 			}
 
-			public int getHarvestLevel() {
+			public int getLevel() {
 				return 1;
 			}
 
-			public int getEnchantability() {
+			public int getEnchantmentValue() {
 				return 1;
 			}
 
-			public Ingredient getRepairMaterial() {
+			public Ingredient getRepairIngredient() {
 				return Ingredient.EMPTY;
 			}
-		}, 3, -3f, new Item.Properties().group(NoobTabItemGroup.tab)) {
-			@Override
-			public boolean hitEntity(ItemStack itemstack, LivingEntity entity, LivingEntity sourceentity) {
-				boolean retval = super.hitEntity(itemstack, entity, sourceentity);
-				double x = entity.getPosX();
-				double y = entity.getPosY();
-				double z = entity.getPosZ();
-				World world = entity.world;
-				{
-					Map<String, Object> $_dependencies = new HashMap<>();
-					$_dependencies.put("entity", entity);
-					$_dependencies.put("itemstack", itemstack);
-					MudAxeMobIsHitWithToolProcedure.executeProcedure($_dependencies);
-				}
-				return retval;
-			}
-		}.setRegistryName("mud_sword"));
+		}, 3, -3f, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
+	}
+
+	@Override
+	public boolean hurtEnemy(ItemStack itemstack, LivingEntity entity, LivingEntity sourceentity) {
+		boolean retval = super.hurtEnemy(itemstack, entity, sourceentity);
+		MudAxeMobIsHitWithToolProcedure.execute(entity, itemstack);
+		return retval;
 	}
 }
