@@ -1,52 +1,41 @@
 
 package io.itch.awesomekalin.noob.item;
 
-import net.minecraftforge.registries.ObjectHolder;
+import net.minecraft.world.item.crafting.Ingredient;
+import net.minecraft.world.item.Tier;
+import net.minecraft.world.item.PickaxeItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Item;
 
-import net.minecraft.item.crafting.Ingredient;
-import net.minecraft.item.PickaxeItem;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Item;
-import net.minecraft.item.IItemTier;
+import io.itch.awesomekalin.noob.init.NoobModTabs;
+import io.itch.awesomekalin.noob.init.NoobModItems;
 
-import io.itch.awesomekalin.noob.itemgroup.NoobTabItemGroup;
-import io.itch.awesomekalin.noob.NoobModElements;
-
-@NoobModElements.ModElement.Tag
-public class NoobPickaxeItem extends NoobModElements.ModElement {
-	@ObjectHolder("noob:noob_pickaxe")
-	public static final Item block = null;
-	public NoobPickaxeItem(NoobModElements instance) {
-		super(instance, 23);
-	}
-
-	@Override
-	public void initElements() {
-		elements.items.add(() -> new PickaxeItem(new IItemTier() {
-			public int getMaxUses() {
+public class NoobPickaxeItem extends PickaxeItem {
+	public NoobPickaxeItem() {
+		super(new Tier() {
+			public int getUses() {
 				return 46;
 			}
 
-			public float getEfficiency() {
+			public float getSpeed() {
 				return 3f;
 			}
 
-			public float getAttackDamage() {
+			public float getAttackDamageBonus() {
 				return -1f;
 			}
 
-			public int getHarvestLevel() {
+			public int getLevel() {
 				return 1;
 			}
 
-			public int getEnchantability() {
+			public int getEnchantmentValue() {
 				return 4;
 			}
 
-			public Ingredient getRepairMaterial() {
-				return Ingredient.fromStacks(new ItemStack(NoobDustItem.block));
+			public Ingredient getRepairIngredient() {
+				return Ingredient.of(new ItemStack(NoobModItems.NOOB_DUST.get()));
 			}
-		}, 1, -3f, new Item.Properties().group(NoobTabItemGroup.tab)) {
-		}.setRegistryName("noob_pickaxe"));
+		}, 1, -3f, new Item.Properties().tab(NoobModTabs.TAB_NOOB_TAB));
 	}
 }
