@@ -1,248 +1,70 @@
 
 package io.itch.awesomekalin.noob.world.biome;
 
-import net.minecraftforge.registries.ObjectHolder;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.api.distmarker.Dist;
+import net.minecraft.world.level.levelgen.placement.SurfaceWaterDepthFilter;
+import net.minecraft.world.level.levelgen.placement.RarityFilter;
+import net.minecraft.world.level.levelgen.placement.NoiseThresholdCountPlacement;
+import net.minecraft.world.level.levelgen.placement.InSquarePlacement;
+import net.minecraft.world.level.levelgen.placement.CountPlacement;
+import net.minecraft.world.level.levelgen.placement.BiomeFilter;
+import net.minecraft.world.level.levelgen.feature.trunkplacers.StraightTrunkPlacer;
+import net.minecraft.world.level.levelgen.feature.stateproviders.BlockStateProvider;
+import net.minecraft.world.level.levelgen.feature.foliageplacers.BlobFoliagePlacer;
+import net.minecraft.world.level.levelgen.feature.featuresize.TwoLayersFeatureSize;
+import net.minecraft.world.level.levelgen.feature.configurations.TreeConfiguration;
+import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.GenerationStep;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.biome.MobSpawnSettings;
+import net.minecraft.world.level.biome.BiomeSpecialEffects;
+import net.minecraft.world.level.biome.BiomeGenerationSettings;
+import net.minecraft.world.level.biome.Biome;
+import net.minecraft.util.valueproviders.ConstantInt;
+import net.minecraft.data.worldgen.placement.PlacementUtils;
+import net.minecraft.data.worldgen.features.VegetationFeatures;
+import net.minecraft.data.worldgen.features.FeatureUtils;
+import net.minecraft.data.worldgen.BiomeDefaultFeatures;
 
-import net.minecraft.world.gen.surfacebuilders.SurfaceBuilderConfig;
-import net.minecraft.world.gen.surfacebuilders.SurfaceBuilder;
-import net.minecraft.world.gen.placement.Placement;
-import net.minecraft.world.gen.placement.FrequencyConfig;
-import net.minecraft.world.gen.placement.AtSurfaceWithExtraConfig;
-import net.minecraft.world.gen.feature.structure.MineshaftStructure;
-import net.minecraft.world.gen.feature.structure.MineshaftConfig;
-import net.minecraft.world.gen.feature.IFeatureConfig;
-import net.minecraft.world.gen.feature.Feature;
-import net.minecraft.world.gen.feature.BaseTreeFeatureConfig;
-import net.minecraft.world.gen.feature.AbstractTreeFeature;
-import net.minecraft.world.gen.blockstateprovider.SimpleBlockStateProvider;
-import net.minecraft.world.gen.IWorldGenerationReader;
-import net.minecraft.world.gen.GenerationStage;
-import net.minecraft.world.biome.DefaultBiomeFeatures;
-import net.minecraft.world.biome.Biome;
-import net.minecraft.world.IWorldWriter;
-import net.minecraft.world.IWorld;
-import net.minecraft.util.math.MutableBoundingBox;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.Direction;
-import net.minecraft.tags.BlockTags;
-import net.minecraft.block.material.Material;
-import net.minecraft.block.Blocks;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.Block;
+import java.util.List;
 
-import java.util.Set;
-import java.util.Random;
+import io.itch.awesomekalin.noob.world.features.treedecorators.NoobBioTrunkDecorator;
+import io.itch.awesomekalin.noob.world.features.treedecorators.NoobBioLeaveDecorator;
+import io.itch.awesomekalin.noob.world.features.treedecorators.NoobBioFruitDecorator;
 
-import io.itch.awesomekalin.noob.NoobModElements;
+import com.google.common.collect.ImmutableList;
 
-@NoobModElements.ModElement.Tag
-public class NoobBioBiome extends NoobModElements.ModElement {
-	@ObjectHolder("noob:noob_bio")
-	public static final CustomBiome biome = null;
-	public NoobBioBiome(NoobModElements instance) {
-		super(instance, 5);
+public class NoobBioBiome {
+	public static Biome createBiome() {
+		BiomeSpecialEffects effects = new BiomeSpecialEffects.Builder().fogColor(-5916161).waterColor(-14329397).waterFogColor(-14329397)
+				.skyColor(-5916161).foliageColorOverride(-13261999).grassColorOverride(-13261999).build();
+		BiomeGenerationSettings.Builder biomeGenerationSettings = new BiomeGenerationSettings.Builder();
+		biomeGenerationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION,
+				PlacementUtils.register("noob:tree_noob_bio",
+						FeatureUtils.register("noob:tree_noob_bio", Feature.TREE,
+								new TreeConfiguration.TreeConfigurationBuilder(BlockStateProvider.simple(Blocks.DIRT.defaultBlockState()),
+										new StraightTrunkPlacer(5, 2, 0), BlockStateProvider.simple(Blocks.DIRT.defaultBlockState()),
+										new BlobFoliagePlacer(ConstantInt.of(2), ConstantInt.of(0), 3), new TwoLayersFeatureSize(1, 0, 1))
+										.decorators(ImmutableList.of(NoobBioLeaveDecorator.INSTANCE, NoobBioTrunkDecorator.INSTANCE,
+												NoobBioFruitDecorator.INSTANCE))
+										.build()),
+						List.of(CountPlacement.of(4), InSquarePlacement.spread(), SurfaceWaterDepthFilter.forMaxDepth(0),
+								PlacementUtils.HEIGHTMAP_OCEAN_FLOOR, PlacementUtils.filteredByBlockSurvival(Blocks.OAK_SAPLING),
+								BiomeFilter.biome())));
+		biomeGenerationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION,
+				PlacementUtils.register("noob:grass_noob_bio", VegetationFeatures.PATCH_GRASS, List.of(NoiseThresholdCountPlacement.of(-0.8D, 5, 1),
+						InSquarePlacement.spread(), PlacementUtils.HEIGHTMAP_WORLD_SURFACE, BiomeFilter.biome())));
+		biomeGenerationSettings.addFeature(GenerationStep.Decoration.VEGETAL_DECORATION,
+				PlacementUtils.register("noob:flower_noob_bio", VegetationFeatures.FLOWER_DEFAULT, List.of(CountPlacement.of(1),
+						RarityFilter.onAverageOnceEvery(32), InSquarePlacement.spread(), PlacementUtils.HEIGHTMAP, BiomeFilter.biome())));
+		BiomeDefaultFeatures.addDefaultCarversAndLakes(biomeGenerationSettings);
+		BiomeDefaultFeatures.addDefaultOres(biomeGenerationSettings);
+		BiomeDefaultFeatures.addSurfaceFreezing(biomeGenerationSettings);
+		BiomeDefaultFeatures.addDefaultMonsterRoom(biomeGenerationSettings);
+		MobSpawnSettings.Builder mobSpawnInfo = new MobSpawnSettings.Builder();
+		return new Biome.BiomeBuilder().precipitation(Biome.Precipitation.RAIN).biomeCategory(Biome.BiomeCategory.NONE).temperature(0.5f).downfall(1f)
+				.specialEffects(effects).mobSpawnSettings(mobSpawnInfo.build()).generationSettings(biomeGenerationSettings.build()).build();
 	}
 
-	@Override
-	public void initElements() {
-		elements.biomes.add(() -> new CustomBiome());
-	}
-
-	@Override
-	public void init(FMLCommonSetupEvent event) {
-	}
-	static class CustomBiome extends Biome {
-		public CustomBiome() {
-			super(new Biome.Builder().downfall(1f).depth(0.1f).scale(0.2f).temperature(0.5f).precipitation(Biome.RainType.RAIN)
-					.category(Biome.Category.NONE).waterColor(-14329397).waterFogColor(-14329397).surfaceBuilder(SurfaceBuilder.DEFAULT,
-							new SurfaceBuilderConfig(Blocks.DIRT.getDefaultState(), Blocks.DIRT.getDefaultState(), Blocks.DIRT.getDefaultState())));
-			setRegistryName("noob_bio");
-			DefaultBiomeFeatures.addCarvers(this);
-			DefaultBiomeFeatures.addMonsterRooms(this);
-			DefaultBiomeFeatures.addOres(this);
-			DefaultBiomeFeatures.addLakes(this);
-			DefaultBiomeFeatures.addFreezeTopLayer(this);
-			this.addStructure(Feature.STRONGHOLD.withConfiguration(IFeatureConfig.NO_FEATURE_CONFIG));
-			this.addStructure(Feature.MINESHAFT.withConfiguration(new MineshaftConfig(0.004D, MineshaftStructure.Type.NORMAL)));
-			this.addStructure(Feature.PILLAGER_OUTPOST.withConfiguration(IFeatureConfig.NO_FEATURE_CONFIG));
-			addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, Feature.FLOWER.withConfiguration(DefaultBiomeFeatures.DEFAULT_FLOWER_CONFIG)
-					.withPlacement(Placement.COUNT_HEIGHTMAP_32.configure(new FrequencyConfig(1))));
-			addFeature(GenerationStage.Decoration.VEGETAL_DECORATION, Feature.RANDOM_PATCH.withConfiguration(DefaultBiomeFeatures.GRASS_CONFIG)
-					.withPlacement(Placement.COUNT_HEIGHTMAP_DOUBLE.configure(new FrequencyConfig(1))));
-			addFeature(GenerationStage.Decoration.VEGETAL_DECORATION,
-					new CustomTreeFeature()
-							.withConfiguration((new BaseTreeFeatureConfig.Builder(new SimpleBlockStateProvider(Blocks.DIRT.getDefaultState()),
-									new SimpleBlockStateProvider(Blocks.DIRT.getDefaultState()))).baseHeight(5)
-											.setSapling((net.minecraftforge.common.IPlantable) Blocks.JUNGLE_SAPLING).build())
-							.withPlacement(Placement.COUNT_EXTRA_HEIGHTMAP.configure(new AtSurfaceWithExtraConfig(4, 0.1F, 1))));
-		}
-
-		@OnlyIn(Dist.CLIENT)
-		@Override
-		public int getGrassColor(double posX, double posZ) {
-			return -13261999;
-		}
-
-		@OnlyIn(Dist.CLIENT)
-		@Override
-		public int getFoliageColor() {
-			return -13261999;
-		}
-
-		@OnlyIn(Dist.CLIENT)
-		@Override
-		public int getSkyColor() {
-			return -5916161;
-		}
-	}
-
-	static class CustomTreeFeature extends AbstractTreeFeature<BaseTreeFeatureConfig> {
-		CustomTreeFeature() {
-			super(BaseTreeFeatureConfig::deserialize);
-		}
-
-		@Override
-		protected boolean place(IWorldGenerationReader worldgen, Random rand, BlockPos position, Set<BlockPos> changedBlocks,
-				Set<BlockPos> changedBlocks2, MutableBoundingBox bbox, BaseTreeFeatureConfig conf) {
-			if (!(worldgen instanceof IWorld))
-				return false;
-			IWorld world = (IWorld) worldgen;
-			int height = rand.nextInt(5) + 5;
-			boolean spawnTree = true;
-			if (position.getY() >= 1 && position.getY() + height + 1 <= world.getHeight()) {
-				for (int j = position.getY(); j <= position.getY() + 1 + height; j++) {
-					int k = 1;
-					if (j == position.getY())
-						k = 0;
-					if (j >= position.getY() + height - 1)
-						k = 2;
-					for (int px = position.getX() - k; px <= position.getX() + k && spawnTree; px++) {
-						for (int pz = position.getZ() - k; pz <= position.getZ() + k && spawnTree; pz++) {
-							if (j >= 0 && j < world.getHeight()) {
-								if (!this.isReplaceable(world, new BlockPos(px, j, pz))) {
-									spawnTree = false;
-								}
-							} else {
-								spawnTree = false;
-							}
-						}
-					}
-				}
-				if (!spawnTree) {
-					return false;
-				} else {
-					Block ground = world.getBlockState(position.add(0, -1, 0)).getBlock();
-					Block ground2 = world.getBlockState(position.add(0, -2, 0)).getBlock();
-					if (!((ground == Blocks.DIRT || ground == Blocks.DIRT) && (ground2 == Blocks.DIRT || ground2 == Blocks.DIRT)))
-						return false;
-					BlockState state = world.getBlockState(position.down());
-					if (position.getY() < world.getHeight() - height - 1) {
-						setTreeBlockState(changedBlocks, world, position.down(), Blocks.DIRT.getDefaultState(), bbox);
-						for (int genh = position.getY() - 3 + height; genh <= position.getY() + height; genh++) {
-							int i4 = genh - (position.getY() + height);
-							int j1 = (int) (1 - i4 * 0.5);
-							for (int k1 = position.getX() - j1; k1 <= position.getX() + j1; ++k1) {
-								for (int i2 = position.getZ() - j1; i2 <= position.getZ() + j1; ++i2) {
-									int j2 = i2 - position.getZ();
-									if (Math.abs(position.getX()) != j1 || Math.abs(j2) != j1 || rand.nextInt(2) != 0 && i4 != 0) {
-										BlockPos blockpos = new BlockPos(k1, genh, i2);
-										state = world.getBlockState(blockpos);
-										if (state.getBlock().isAir(state, world, blockpos) || state.getMaterial().blocksMovement()
-												|| state.isIn(BlockTags.LEAVES) || state.getBlock() == Blocks.DIRT
-												|| state.getBlock() == Blocks.DIRT) {
-											setTreeBlockState(changedBlocks, world, blockpos, Blocks.DIRT.getDefaultState(), bbox);
-										}
-									}
-								}
-							}
-						}
-						for (int genh = 0; genh < height; genh++) {
-							BlockPos genhPos = position.up(genh);
-							state = world.getBlockState(genhPos);
-							setTreeBlockState(changedBlocks, world, genhPos, Blocks.DIRT.getDefaultState(), bbox);
-							if (state.getBlock().isAir(state, world, genhPos) || state.getMaterial().blocksMovement() || state.isIn(BlockTags.LEAVES)
-									|| state.getBlock() == Blocks.DIRT || state.getBlock() == Blocks.DIRT) {
-								if (genh > 0) {
-									if (rand.nextInt(3) > 0 && world.isAirBlock(position.add(-1, genh, 0)))
-										setTreeBlockState(changedBlocks, world, position.add(-1, genh, 0), Blocks.DIRT.getDefaultState(), bbox);
-									if (rand.nextInt(3) > 0 && world.isAirBlock(position.add(1, genh, 0)))
-										setTreeBlockState(changedBlocks, world, position.add(1, genh, 0), Blocks.DIRT.getDefaultState(), bbox);
-									if (rand.nextInt(3) > 0 && world.isAirBlock(position.add(0, genh, -1)))
-										setTreeBlockState(changedBlocks, world, position.add(0, genh, -1), Blocks.DIRT.getDefaultState(), bbox);
-									if (rand.nextInt(3) > 0 && world.isAirBlock(position.add(0, genh, 1)))
-										setTreeBlockState(changedBlocks, world, position.add(0, genh, 1), Blocks.DIRT.getDefaultState(), bbox);
-								}
-							}
-						}
-						for (int genh = position.getY() - 3 + height; genh <= position.getY() + height; genh++) {
-							int k4 = (int) (1 - (genh - (position.getY() + height)) * 0.5);
-							for (int genx = position.getX() - k4; genx <= position.getX() + k4; genx++) {
-								for (int genz = position.getZ() - k4; genz <= position.getZ() + k4; genz++) {
-									BlockPos bpos = new BlockPos(genx, genh, genz);
-									state = world.getBlockState(bpos);
-									if (state.isIn(BlockTags.LEAVES) || state.getBlock() == Blocks.DIRT) {
-										BlockPos blockpos1 = bpos.south();
-										BlockPos blockpos2 = bpos.west();
-										BlockPos blockpos3 = bpos.east();
-										BlockPos blockpos4 = bpos.north();
-										if (rand.nextInt(4) == 0 && world.isAirBlock(blockpos2))
-											this.addVines(world, blockpos2, changedBlocks, bbox);
-										if (rand.nextInt(4) == 0 && world.isAirBlock(blockpos3))
-											this.addVines(world, blockpos3, changedBlocks, bbox);
-										if (rand.nextInt(4) == 0 && world.isAirBlock(blockpos4))
-											this.addVines(world, blockpos4, changedBlocks, bbox);
-										if (rand.nextInt(4) == 0 && world.isAirBlock(blockpos1))
-											this.addVines(world, blockpos1, changedBlocks, bbox);
-									}
-								}
-							}
-						}
-						if (rand.nextInt(4) == 0 && height > 5) {
-							for (int hlevel = 0; hlevel < 2; hlevel++) {
-								for (Direction Direction : Direction.Plane.HORIZONTAL) {
-									if (rand.nextInt(4 - hlevel) == 0) {
-										Direction dir = Direction.getOpposite();
-										setTreeBlockState(changedBlocks, world, position.add(dir.getXOffset(), height - 5 + hlevel, dir.getZOffset()),
-												Blocks.DIRT.getDefaultState(), bbox);
-									}
-								}
-							}
-						}
-						return true;
-					} else {
-						return false;
-					}
-				}
-			} else {
-				return false;
-			}
-		}
-
-		private void addVines(IWorld world, BlockPos pos, Set<BlockPos> changedBlocks, MutableBoundingBox bbox) {
-			setTreeBlockState(changedBlocks, world, pos, Blocks.DIRT.getDefaultState(), bbox);
-			int i = 5;
-			for (BlockPos blockpos = pos.down(); world.isAirBlock(blockpos) && i > 0; --i) {
-				setTreeBlockState(changedBlocks, world, blockpos, Blocks.DIRT.getDefaultState(), bbox);
-				blockpos = blockpos.down();
-			}
-		}
-
-		private boolean canGrowInto(Block blockType) {
-			return blockType.getDefaultState().getMaterial() == Material.AIR || blockType == Blocks.DIRT || blockType == Blocks.DIRT
-					|| blockType == Blocks.DIRT || blockType == Blocks.DIRT;
-		}
-
-		private boolean isReplaceable(IWorld world, BlockPos pos) {
-			BlockState state = world.getBlockState(pos);
-			return state.getBlock().isAir(state, world, pos) || canGrowInto(state.getBlock()) || !state.getMaterial().blocksMovement();
-		}
-
-		private void setTreeBlockState(Set<BlockPos> changedBlocks, IWorldWriter world, BlockPos pos, BlockState state, MutableBoundingBox mbb) {
-			super.func_227217_a_(world, pos, state, mbb);
-			changedBlocks.add(pos.toImmutable());
-		}
+	public static void init() {
 	}
 }
